Rename MonitoringSubscriber file for autoloading
The non-zeitwerk friendly naming was causing a Jenkins build failure when
creating the mongo indexes.
