+++ app/lib/mongo_instrumentation.rb
@@ -1,3 +1,3 @@
 require "mongo_instrumentation/runtime_registry"
-require "mongo_instrumentation/moped_subscriber"
+require "mongo_instrumentation/monitoring_subscriber"
 require "mongo_instrumentation/controller_runtime"
