+++ app/lib/mongo_instrumentation/monitoring_subscriber.rb
